% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_collaborators.R
\name{add_qfieldcloud_collaborator}
\alias{add_qfieldcloud_collaborator}
\title{Add a collaborator to a QFieldCloud project}
\usage{
add_qfieldcloud_collaborator(token, endpoint, project_id, collaborator, role)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}

\item{collaborator}{QFieldCloud username to add}

\item{role}{QFieldCloud role to assign the user within the project}
}
\value{
string indicating success or failure
}
\description{
Add a collaborator to a QFieldCloud project
}
