% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_collaborators.R
\name{delete_qfieldcloud_collaborator}
\alias{delete_qfieldcloud_collaborator}
\title{Delete a collaborator to a QFieldCloud project}
\usage{
delete_qfieldcloud_collaborator(token, endpoint, project_id, username)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}

\item{username}{QFieldCloud username to delete}
}
\value{
string indicating success or failure
}
\description{
Delete a collaborator to a QFieldCloud project
}
