% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_collaborators.R
\name{get_qfieldcloud_collaborator}
\alias{get_qfieldcloud_collaborator}
\title{Get the role of a collaborator in a QFieldCloud project}
\usage{
get_qfieldcloud_collaborator(token, endpoint, project_id, username)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}

\item{username}{QFieldCloud username}
}
\value{
data.frame with two columns storing username and role
}
\description{
Get the role of a collaborator in a QFieldCloud project
}
