% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_collaborators.R
\name{get_qfieldcloud_collaborators}
\alias{get_qfieldcloud_collaborators}
\title{Get collaborators in a QFieldCloud project}
\usage{
get_qfieldcloud_collaborators(token, endpoint, project_id)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}
}
\value{
data.frame with two columns storing project names and project ids
}
\description{
Get collaborators in a QFieldCloud project
}
