% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_files.R
\name{get_qfieldcloud_file}
\alias{get_qfieldcloud_file}
\title{Get QFieldCloud file}
\usage{
get_qfieldcloud_file(token, endpoint, project_id, filename)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{project id corresponding to project to download file from}

\item{filename}{filename of file to download}
}
\value{
dataframe with filename and path to temporary location where file is downloaded
}
\description{
Get QFieldCloud file
}
