% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_files.R
\name{get_qfieldcloud_files}
\alias{get_qfieldcloud_files}
\title{Get QFieldCloud files in project (GeoPackages only)}
\usage{
get_qfieldcloud_files(token, endpoint, project_id)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{project id corresponding to project to download file from}
}
\value{
data.frame with two columns storing file name and last modified data
}
\description{
Get QFieldCloud files in project (GeoPackages only)
}
