% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_projects.R
\name{get_qfieldcloud_project}
\alias{get_qfieldcloud_project}
\title{Get a QFieldCloud project}
\usage{
get_qfieldcloud_project(token, endpoint, project_id)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}
}
\value{
data.frame with project information
}
\description{
Get a QFieldCloud project
}
