% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_projects.R
\name{get_qfieldcloud_projects}
\alias{get_qfieldcloud_projects}
\title{Get QFieldCloud projects for authenticated user}
\usage{
get_qfieldcloud_projects(token, endpoint)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}
}
\value{
data.frame with three columns storing project names, project ids, and user roles
}
\description{
Get QFieldCloud projects for authenticated user
}
