% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_users.R
\name{get_qfieldcloud_users}
\alias{get_qfieldcloud_users}
\title{Get QFieldCloud users}
\usage{
get_qfieldcloud_users(token, endpoint)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}
}
\value{
data.frame with two columns storing usernames and full names.
}
\description{
Get QFieldCloud users
}
