% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_files.R
\name{post_qfieldcloud_file}
\alias{post_qfieldcloud_file}
\title{Create QFieldCloud files}
\usage{
post_qfieldcloud_file(token, endpoint, project_id, filename, file_path)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}

\item{filename}{QFieldCloud filename}

\item{file_path}{file path of file to upload to QFieldCloud project}
}
\value{
string indicating project creation success or failure
}
\description{
Create QFieldCloud files
}
