% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_projects.R
\name{post_qfieldcloud_project}
\alias{post_qfieldcloud_project}
\title{Create a QFieldCloud project}
\usage{
post_qfieldcloud_project(token, endpoint, project_name)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_name}{QFieldCloud project name}
}
\value{
string indicating project creation success or failure
}
\description{
Create a QFieldCloud project
}
