% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfieldcloud_collaborators.R
\name{update_qfieldcloud_collaborator}
\alias{update_qfieldcloud_collaborator}
\title{Update a collaborator to a QFieldCloud project}
\usage{
update_qfieldcloud_collaborator(token, endpoint, project_id, username, role)
}
\arguments{
\item{token}{session token}

\item{endpoint}{QFieldCloud app url (omit https:// and trailing /)}

\item{project_id}{QFieldCloud project id}

\item{username}{QFieldCloud username to add}

\item{role}{QFieldCloud role to assign the user within the project}
}
\value{
string indicating success or failure
}
\description{
Update a collaborator to a QFieldCloud project
}
